Backend/requirements.txt
accelerate==0.20.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
banks==2.1.2
bitsandbytes==0.40.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.7.4
filelock==3.18.0
filetype==1.2.0
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.16.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
langdetect==1.0.9
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index-core==0.12.34.post1
llama-parse==0.6.21
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==1.25.2
openai==1.77.0
packaging==25.0
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.22.5
PyPDF2==3.0.1
pytest==7.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.4.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
services==0.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Wand==0.6.13
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.0